move ticker to schedule
